Refactor test module
